[238414]  Merge simon_pre_0_9 to HEAD https://bugs.eclipse.org/bugs/show_bug.cgi?id=238414

+++ plugins/org.eclipse.emf.cdo.server/src/org/eclipse/emf/cdo/internal/server/QueryResult.java
@@ -20,7 +20,7 @@ import org.eclipse.emf.cdo.server.IView;
  */
 public class QueryResult extends AbstractQueryResult<Object>
 {
-  public QueryResult(IView view, CDOQueryInfo parameter, long queryID)
+  public QueryResult(IView view, CDOQueryInfo parameter, int queryID)
   {
     super(view, parameter, queryID);
   }
